Add a python wrapper to mcumgr-client
This commit adds a python module invoking mcumgr-client binary and the
CI configuration to build wheels in various environment.

To improve portability of wheels, default-features are disabled for
serialport as it removes the dependency on libudev. There are no
drawbacks that I know of as libudev is used to enumerate serial ports,
which mcumgr-client does not use.

Signed-off-by: Adrien Bruant <adrien.bruant@aalberts-hfc.com>