Split mcumgr-client into lib and bin crates
In order to provide integration with other programs (such as python
bindings), the core functionalities of this library are refactored and
exposed in lib.rs.

All the functionalities related to the command line interface are moved
to main.rs.

The refactor includes:
- parameters related to the configuration of the serial port are moved
to dedicated struct SerialSpecs, built `From` the Cli structure,
- move cli option '--slot' to subcommand 'upload',
- list() now returns a serde_cbor::Value,
- upload() takes a FnMut callback to inform on the progress of the
operation.

Note that the compilation of the library now generates warning as many
structs defined in nmp_hdr were never used in the first place.